requirements.txt
django-environ==0.8.1
mysqlclient==2.1.0
Django==3.2.9
django-mptt==0.13.4
django-ckeditor==6.2.0
django-debug-toolbar==3.2.3
six==1.16.0
future==0.18.2
django-constance==2.8.0
django-picklefield==3.0.1
sorl-thumbnail==12.7.0
Pillow==8.4.0